fixing warnings.
ioeye-staging-commit :: 45f2f518f5d12f31c1302165e3e7b38e8537449b

+++ common/instamsg/driver/sg_stdlib.c
@@ -112,10 +112,10 @@ char* sg_strnstr(const char *str1, const char *str2, int maxSize)
         return NULL;
     }
 
-    for(i = 0; i <= (maxSize - strlen(str2)); i++)
+    for(i = 0; i <= (maxSize - ((int)(strlen(str2)))); i++)
     {
         unsigned char failure = 0;
-        for(j = 0; j < strlen(str2); j++)
+        for(j = 0; j < ((int)(strlen(str2))); j++)
         {
             if(str1[i + j] != str2[j])
             {
